Fix authentication glitch for LDAP users where user_dn and/or uid_attr were not expressly set
auth_LDAP was using isset to check authparams values, which was causing some logic issues if those values had been unset (ie. were "" instead of NULL).  Changed them over to use PHP's empty function (and inverted the corresponding if statements) and this seems to have solved the issue and allows the code to fall to its already defined defaults if those fields are empty.