fix LDAP user lookups for non-NetIQ types
two issues to address:
- the $type variable was not valid for the info() function
- the fallbacks were AD specific

so an authentication method of type "LDAP" would result in a
filter of something like:
(&(objectCategory=person)(samaccountname=parsonsa*))

when it should be like this for OpenLDAP or FreeIPA (389):
(&(objectClass=person)(uid=parsonsa*))

I've tested against FreeIPA, which is my use case.  this should
not alter the behavior for AD, which is likely the most common
use case.