Bugfix: Sections with "." are excluded when performing an export. Fixes #2260
Pass requested sections for export by 'id' value instead of by 'name'.

http://php.net/manual/en/language.variables.external.php

Typically, PHP does not alter the names of variables when they are passed into a script.
However, it should be noted that the dot (period, full stop) is not a valid character in a PHP variable name.
PHP will automatically replace any dots in incoming variable names with underscores