l10n update - making the installer wizard translatable (#3248)
* l10n: make installer wizards translatable

Rework the existing files in app/install so all the possible generated
text output is passed through the l10n macro _() to get all strings
localized.

Some existing parts of sentences within the install wizards got
rephrased and reworked so localized parts of the concatenated strings
are orthographically correct writeable in the target language.

The localization is currently not working yet as the UI isn't using the
locale setting from the browser e.g., the situation is the same as in
the login site. phpIPAM is using the language for the user from the
setting within the database for the user and while the installer is
running there is no access to the database given.

* l10n: Update German translations

Update the po file and also the binary message catalogue for German after
reworking various source files.