l10n update - small various strings added or reworked (#3260)
* l10n: Make status messages in Circuits translatable

Make some entries for the indication of the status from a circuit
translatable within the main overview for circuits and the edit page for
a circuit.

* l10n: Add more strings in some import/export dialogs

A few more strings in two dialogs which can be added to the localization
catalogue.

* l10n: Modify localization function for password string

The previous usage of _($p) wasn't working as the strings of the used
array 'types' aren't localized anywhere yet.
Adding the two used strings to the l10n catalogue and modify the string
concatenation so the generation is localized.

* l10n: Adding more various translatable strings

* l10n: Update German translations

Update the po file and also the binary message catalogue for German after
reworking various source files.

Closes #3260